style/Changed card style
