+++ client/src/components/Card/_card.scss
@@ -1,21 +1,24 @@
-
 .card {
-    display: flex;
-    flex-direction: column;
-    justify-content: space-between; /* push card content to the top and bottom of the card */
-    background-color: rgb(211, 209, 209);
-    border-radius: 10px;
-    
-    box-sizing: border-box;
-    width: calc(100% - 40px);
-    margin: 20px;
-    border: 1px solid rgb(156, 156, 156);
-    box-shadow: 0px 0px 5px 1px rgba(0, 0, 0, 0.185);
-  }
-  
-  .card-content {
-    display: flex;
-    flex: 1; /* take up all available vertical space in the card */
-    flex-direction: column;
-    justify-content: flex-end; /* push accordion to the bottom of the card content */
-  }
+  width: 250px;
+  display: flex;
+  flex-direction: column;
+  margin: 22px;
+  background-color: #eee9e9;
+  border: 2px dashed rgb(87, 86, 86);
+  color: rgb(87, 86, 86);
+  border-radius: 6px;
+}
+
+.card-header {
+  padding: 10px 10px 0px 10px;
+  margin-bottom: 10px;
+}
+
+.card-content {
+  display: flex;
+  flex: 1;
+  /* take up all available vertical space in the card */
+  flex-direction: column;
+  justify-content: flex-end;
+  /* push accordion to the bottom of the card content */
+}
