+++ client/src/components/PopUpComponent/_PopUp.scss
@@ -6,65 +6,4 @@
     width: 250px;
     display: flex;
     flex-direction: column;
-   
-   
-
-    p {
-        margin: 0px 0px 0px 0px !important;
-        font-size: 0.88rem;
-    }
-
-    .header {
-        display: flex;
-        flex-direction: row;
-        justify-content: space-between;
-        margin-bottom: 5px;
-
-
-        h2 {
-            font-weight: 600;
-            font-size: 1.2rem;
-            text-transform: uppercase;
-            margin: 0px 15px 0px 0px !important;
-        }
-
-
-    }
-
-
-    .body {
-        display: flex;
-        flex-direction: row;
-        justify-content: space-between;
-
-        margin-bottom: 8px;
-
-        .state {
-            margin: 0px 5px 0px 0px !important;
-        }
-
-        .thisDate {
-            border-radius: 6px;
-            padding-left: 5px;
-            padding-right: 5px;
-
-            background-color: rgb(211, 208, 208);
-        }
-
-    }
-
-
-    .position {
-        display: flex;
-        flex-direction: row;
-        justify-content: space-between;
-
-        p {
-            border-radius: 6px;
-            padding-left: 5px;
-            padding-right: 5px;
-
-            background-color: rgb(211, 208, 208);
-        }
-    }
 }
