+++ client/src/components/Sidebar/Sidebar.tsx
@@ -8,7 +8,7 @@ interface SidebarProps {
 }
 
 export const Sidebar: React.FC<SidebarProps> = ({equipments}) => {
-  const componentsArray = [<Card model={equipments[0].typeId} name={equipments[0].equipName} date={equipments[0].getMostRecentDate()} position={[1,2]} state={"03b2d446-e3ba-4c82-8dc2-a5611fea6e1f"} />];
+  const componentsArray = [<Card equipment={equipments[0]}  />];
   return (
     <div className="sidebar" >
       <Search />
