+++ client/src/components/Sidebar/_sidebar.scss
@@ -1,7 +1,7 @@
 .sidebar {
     height: 100%;
     width: 100%;
-    background-color: rgb(209, 208, 208);
+    background-color: rgb(211, 208, 208);
     border-right: 2px solid rgb(156, 156, 156);
     box-sizing: border-box;
     box-shadow: 0 0 10px rgba(0, 0, 0, 0.1);
