+++ client/src/scss/global.scss
@@ -9,6 +9,7 @@
 @import "../components/Checkbox/checkbox";
 @import "../components/PopUpComponent/PopUp";
 @import "../components/Accordeon/Accordeon";
+@import "../components/InnerCard/innerCard";
 
 body {
   font-family: 'Nunito', sans-serif;
